+++ mitra_models/src/notifications/helpers.rs
@@ -35,6 +35,7 @@ pub async fn create_follow_notification(
         None,
         None,
         None,
+        None,
         EventType::Follow,
     ).await
 }
@@ -56,6 +57,7 @@ pub async fn create_follow_request_notification(
         db_client,
         sender_id,
         recipient_id,
+        None,
         None,
         None,
         None,
@@ -84,6 +86,7 @@ pub async fn create_reply_notification(
         Some(post_id),
         None,
         None,
+        None,
         EventType::Reply,
     ).await
 }
@@ -110,6 +113,7 @@ pub async fn create_reaction_notification(
         Some(post_id),
         Some(reaction_id),
         None,
+        None,
         EventType::Reaction,
     ).await
 }
@@ -133,6 +137,7 @@ pub async fn create_mention_notification(
         sender_id,
         recipient_id,
         Some(post_id),
+        None,
         None,
         None,
         EventType::Mention,
@@ -160,6 +165,7 @@ pub async fn create_repost_notification(
         Some(post_id),
         None,
         None,
+        None,
         EventType::Repost,
     ).await
 }
@@ -177,6 +183,7 @@ pub async fn create_subscriber_payment_notification(
         None,
         None,
         Some(invoice_id),
+        None,
         EventType::SubscriberPayment,
     ).await
 }
@@ -190,6 +197,7 @@ pub async fn create_subscriber_leaving_notification(
         db_client,
         sender_id,
         recipient_id,
+        None,
         None,
         None,
         None,
@@ -209,6 +217,7 @@ pub async fn create_subscription_expiration_notification(
         None,
         None,
         None,
+        None,
         EventType::SubscriptionExpiration,
     ).await
 }
@@ -222,6 +231,7 @@ pub async fn create_move_notification(
         db_client,
         sender_id,
         recipient_id,
+        None,
         None,
         None,
         None,
@@ -242,10 +252,29 @@ pub async fn create_signup_notifications(
             None,
             None,
             None,
+            None,
             EventType::SignUp,
         ).await?;
     };
     Ok(())
+}
+
+pub async fn create_moderation_warning_notification(
+    db_client: &impl DatabaseClient,
+    sender_id: Uuid,
+    recipient_id: Uuid,
+    moderation_action_id: Uuid,
+) -> Result<(), DatabaseError> {
+    create_notification(
+        db_client,
+        sender_id,
+        recipient_id,
+        None,
+        None,
+        None,
+        Some(moderation_action_id),
+        EventType::ModerationWarning,
+    ).await
 }
 
 #[cfg(test)]