+++ mitra_models/src/notifications/types.rs
@@ -3,14 +3,17 @@ use postgres_types::FromSql;
 use tokio_postgres::Row;
 use uuid::Uuid;
 
-use crate::database::{
-    int_enum::{int_enum_from_sql, int_enum_to_sql},
-    DatabaseError,
-    DatabaseTypeError,
+use crate::{
+    database::{
+        int_enum::{int_enum_from_sql, int_enum_to_sql},
+        DatabaseError,
+        DatabaseTypeError,
+    },
+    emojis::types::CustomEmoji,
+    moderation_actions::types::ModerationAction,
+    posts::types::{Post, PostDetailed},
+    profiles::types::DbActorProfile,
 };
-use crate::emojis::types::CustomEmoji;
-use crate::posts::types::{Post, PostDetailed};
-use crate::profiles::types::DbActorProfile;
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum EventType {
@@ -26,6 +29,7 @@ pub enum EventType {
     Move,
     SignUp,
     SubscriberLeaving,
+    ModerationWarning,
 }
 
 impl From<EventType> for i16 {
@@ -43,6 +47,7 @@ impl From<EventType> for i16 {
             EventType::Move => 10,
             EventType::SignUp => 11,
             EventType::SubscriberLeaving => 12,
+            EventType::ModerationWarning => 13,
         }
     }
 }
@@ -64,6 +69,7 @@ impl TryFrom<i16> for EventType {
             10 => Self::Move,
             11 => Self::SignUp,
             12 => Self::SubscriberLeaving,
+            13 => Self::ModerationWarning,
             _ => return Err(DatabaseTypeError),
         };
         Ok(event_type)
@@ -83,6 +89,7 @@ struct Notification {
     post_id: Option<Uuid>,
     reaction_id: Option<Uuid>,
     invoice_id: Option<Uuid>,
+    moderation_action_id: Option<Uuid>,
     event_type: EventType,
     created_at: DateTime<Utc>,
 }
@@ -94,12 +101,12 @@ pub struct NotificationDetailed {
     pub reaction_content: Option<String>,
     pub reaction_emoji: Option<CustomEmoji>,
     pub payment_amount: Option<i64>,
+    pub moderation_action: Option<ModerationAction>,
     pub event_type: EventType,
     pub created_at: DateTime<Utc>,
 }
 
 impl TryFrom<&Row> for NotificationDetailed {
-
     type Error = DatabaseError;
 
     fn try_from(row: &Row) -> Result<Self, Self::Error> {
@@ -116,6 +123,13 @@ impl TryFrom<&Row> for NotificationDetailed {
         let maybe_reaction_content = row.try_get("reaction_content")?;
         let maybe_reaction_emoji = row.try_get("reaction_emoji")?;
         let maybe_payment_amount = row.try_get("payment_amount")?;
+        let maybe_moderation_action: Option<ModerationAction> =
+            row.try_get("moderation_action")?;
+        if maybe_moderation_action.as_ref().map(|action| action.id)
+            != db_notification.moderation_action_id
+        {
+            return Err(DatabaseError::type_error());
+        };
         let notification = Self {
             id: db_notification.id,
             sender: db_sender,
@@ -123,6 +137,7 @@ impl TryFrom<&Row> for NotificationDetailed {
             reaction_content: maybe_reaction_content,
             reaction_emoji: maybe_reaction_emoji,
             payment_amount: maybe_payment_amount,
+            moderation_action: maybe_moderation_action,
             event_type: db_notification.event_type,
             created_at: db_notification.created_at,
         };